requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
django-filter==23.5
Pillow==10.1.0
djangorestframework==3.14.0
python-decouple==3.8
